Fixed warnings

+++ amalgamated/jsonlite/jsonlite.c
@@ -583,7 +583,7 @@ next:
         case 0:
             goto done;
         case 1:
-            bits = *c++ << 16;
+            bits = (uint32_t)(*c++ << 16);
             buffer[0] = encode[(bits & 0x00FC0000) >> 18];
             buffer[1] = encode[(bits & 0x0003F000) >> 12];
             buffer[2] = '=';
@@ -591,8 +591,8 @@ next:
             l = c;
             goto write;
         case 2:
-            bits = *c++ << 16;
-            bits |= *c++ << 8;
+            bits = (uint32_t)(*c++ << 16);
+            bits |= (uint32_t)(*c++ << 8);
             buffer[0] = encode[(bits & 0x00FC0000) >> 18];
             buffer[1] = encode[(bits & 0x0003F000) >> 12];
             buffer[2] = encode[(bits & 0x00000FC0) >> 6];
@@ -600,7 +600,7 @@ next:
             l = c;
             goto write;
         default:
-            bits = *c++ << 16;
+            bits = (uint32_t)(*c++ << 16);
             bits |= *c++ << 8;
             bits |= *c++;
             buffer[0] = encode[(bits & 0x00FC0000) >> 18];
